Move power calculation to shared
