+++ packages/shared/src/index.ts
@@ -6,3 +6,4 @@ export * as experiments from "./experiments";
 export * as permissions from "./permissions";
 export * as sdkVersioning from "./sdk-versioning";
 export * as types from "./types";
+export * as power from "./power";
